Ability to edit a text in app was add
